+++ src/main/java/com/shelfeed/backend/domain/book/dto/response/BookDetailResponse.java
@@ -1,4 +1,4 @@
-package com.shelfeed.backend.domain.book.dto.respond;
+package com.shelfeed.backend.domain.book.dto.response;
 
 import com.shelfeed.backend.domain.book.entity.Book;
 import com.shelfeed.backend.domain.library.enums.ReadingStatus;
