+++ src/main/java/com/shelfeed/backend/domain/book/dto/response/BookReviewResponse.java
@@ -1,4 +1,4 @@
-package com.shelfeed.backend.domain.book.dto.respond;
+package com.shelfeed.backend.domain.book.dto.response;
 
 import com.shelfeed.backend.domain.review.entity.Review;
 import lombok.Builder;
