+++ src/main/java/com/shelfeed/backend/domain/library/dto/response/LibraryBookAddResponse.java
@@ -1,4 +1,4 @@
-package com.shelfeed.backend.domain.library.dto.respond;
+package com.shelfeed.backend.domain.library.dto.response;
 
 import com.shelfeed.backend.domain.library.entity.LibraryBook;
 import com.shelfeed.backend.domain.library.enums.ReadingStatus;
@@ -9,16 +9,17 @@ import java.time.LocalDate;
 
 @Getter
 @Builder
-public class LibraryStatusUpdateResponse {
-
+public class LibraryBookAddResponse {
     private Long libraryBookId;
+    private Long bookId;
     private ReadingStatus status;
     private LocalDate startedAt;
     private LocalDate finishedAt;
 
-    public static LibraryStatusUpdateResponse of(LibraryBook lb) {
-        return LibraryStatusUpdateResponse.builder()
+    public static LibraryBookAddResponse of(LibraryBook lb){
+        return LibraryBookAddResponse.builder()
                 .libraryBookId(lb.getLibraryBookId())
+                .bookId(lb.getBook().getBookId())
                 .status(lb.getStatus())
                 .startedAt(lb.getStartedAt())
                 .finishedAt(lb.getFinishedAt())
