+++ src/main/java/com/shelfeed/backend/domain/library/dto/response/UserLibraryResponse.java
@@ -1,4 +1,4 @@
-package com.shelfeed.backend.domain.library.dto.respond;
+package com.shelfeed.backend.domain.library.dto.response;
 
 import com.shelfeed.backend.domain.member.enums.LibraryVisibility;
 import lombok.Builder;
