+++ src/test/java/com/shelfeed/backend/domain/book/BookServiceTest.java
@@ -5,9 +5,9 @@ import com.shelfeed.backend.domain.book.client.dto.AladinItem;
 import com.shelfeed.backend.domain.book.client.dto.AladinSearchResponse;
 import com.shelfeed.backend.domain.book.dto.request.BookReviewSearchRequest;
 import com.shelfeed.backend.domain.book.dto.request.BookSearchRequest;
-import com.shelfeed.backend.domain.book.dto.respond.BookDetailResponse;
-import com.shelfeed.backend.domain.book.dto.respond.BookReviewListResponse;
-import com.shelfeed.backend.domain.book.dto.respond.BookSearchListResponse;
+import com.shelfeed.backend.domain.book.dto.response.BookDetailResponse;
+import com.shelfeed.backend.domain.book.dto.response.BookReviewListResponse;
+import com.shelfeed.backend.domain.book.dto.response.BookSearchListResponse;
 import com.shelfeed.backend.domain.book.entity.Book;
 import com.shelfeed.backend.domain.book.repository.BookRepository;
 import com.shelfeed.backend.domain.book.service.BookService;
