+++ src/test/java/com/shelfeed/backend/domain/library/LibraryServiceTest.java
@@ -4,10 +4,10 @@ import com.shelfeed.backend.domain.book.entity.Book;
 import com.shelfeed.backend.domain.book.repository.BookRepository;
 import com.shelfeed.backend.domain.library.dto.request.LibraryBookAddRequest;
 import com.shelfeed.backend.domain.library.dto.request.LibraryStatusUpdateRequest;
-import com.shelfeed.backend.domain.library.dto.respond.LibraryBookAddResponse;
-import com.shelfeed.backend.domain.library.dto.respond.LibraryBookDetailResponse;
-import com.shelfeed.backend.domain.library.dto.respond.LibraryListResponse;
-import com.shelfeed.backend.domain.library.dto.respond.UserLibraryResponse;
+import com.shelfeed.backend.domain.library.dto.response.LibraryBookAddResponse;
+import com.shelfeed.backend.domain.library.dto.response.LibraryBookDetailResponse;
+import com.shelfeed.backend.domain.library.dto.response.LibraryListResponse;
+import com.shelfeed.backend.domain.library.dto.response.UserLibraryResponse;
 import com.shelfeed.backend.domain.library.entity.LibraryBook;
 import com.shelfeed.backend.domain.library.enums.ReadingStatus;
 import com.shelfeed.backend.domain.library.repository.LibraryRepository;
